chore: conftest created at the lowest level. id naming also changed
